+++ src/main/java/net/silentchaos512/gear/util/Const.java
@@ -160,10 +160,12 @@ public final class Const {
         public static final DataResource<ITrait> SOFT = DataResource.trait("soft");
         public static final DataResource<ITrait> SPOON = DataResource.trait("spoon");
         public static final DataResource<ITrait> STELLAR = DataResource.trait("stellar");
+        public static final DataResource<ITrait> STURDY = DataResource.trait("sturdy");
         public static final DataResource<ITrait> SYNERGISTIC = DataResource.trait("synergistic");
         public static final DataResource<ITrait> TERMINUS = DataResource.trait("terminus");
         public static final DataResource<ITrait> TURTLE = DataResource.trait("turtle");
         public static final DataResource<ITrait> VENOM = DataResource.trait("venom");
+        public static final DataResource<ITrait> VOID_WARD = DataResource.trait("void_ward");
         public static final DataResource<ITrait> VULCAN = DataResource.trait("vulcan");
 
         // Misc constants